Revert "GUI initial implementation"
