+++ .\src\test\java\dream\fcard\util\data\Test.java
@@ -1,0 +1,7 @@
+package dream.fcard.util.data;
+
+public class Test {
+    public static void main(String[] args) {
+        System.out.println("Hello from JavaRunnerTest");
+    }
+}
