+++ src/main/java/dream/fcard/MainApp.java
@@ -11,6 +11,7 @@ import dream.fcard.core.commons.core.Version;
 import dream.fcard.core.commons.exceptions.DataConversionException;
 import dream.fcard.core.commons.util.ConfigUtil;
 import dream.fcard.core.commons.util.StringUtil;
+import dream.fcard.gui.Ui;
 import dream.fcard.gui.UiManager;
 import javafx.application.Application;
 import javafx.stage.Stage;
@@ -24,7 +25,7 @@ public class MainApp extends Application {
 
     private static final Logger logger = LogsCenter.getLogger(MainApp.class);
 
-    protected UiManager ui;
+    protected Ui ui;
     protected Config config;
 
     @Override
