+++ src/main/java/dream/fcard/gui/Ui.java
@@ -1,0 +1,15 @@
+package dream.fcard.gui;
+
+import javafx.stage.Stage;
+
+/**
+ * API of UI component
+ */
+public interface Ui {
+
+    /**
+     * Starts the UI (and the App).
+     */
+    void start(Stage primaryStage);
+
+}
